feat/1248 - extracted the util to query paged results to a util
