+++ src/utils/index.js
@@ -1,2 +1,3 @@
 export { CONSTANTS } from "@kfarranger/middleware";
 export { default as normalizeResults } from './normalizeResults';
+export { default as executePagedQuery } from './executePagedQuery';
